+++ src/main/java/com/example/last/Service/ForReply/ReplyService.java
@@ -5,6 +5,7 @@ import com.example.last.Entity.Post;
 import com.example.last.Entity.Reply;
 import com.example.last.Repository.PostRepository;
 import com.example.last.Repository.ReplyRepository;
+import com.example.last.security.UserDetailsImpl;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -20,17 +21,20 @@ public class ReplyService {
     }
 
 
-    public String Reply(long id, ReplyDto replyDto){
+    public String Reply(long id, ReplyDto replyDto, UserDetailsImpl userDetails){
         Post post1 = postRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("해당 아이디가 존재하지 않습니다.")
         );
-        Reply reply = new Reply(id,replyDto);
+        Reply reply = new Reply(id,replyDto,userDetails);
         repository.save(reply);
-        return reply.getReplyid()+"성공!";
+        return reply.getPostid()+"번 게시글에"+reply.getReplyid()+"번 댓글 작성";
     }
 
-    public String Delete(long id) {
+    public String Delete(long id,UserDetailsImpl userDetails) {
+        if(!repository.getReferenceById(id).getUsername().equals(userDetails.getUsername())){
+            return "본인 댓글만 삭제 가능합니다.";
+        }
         repository.deleteById(id);
-        return "삭제완료";
+        return id+"번 삭제완료";
     }
 }
